+++ volmdlr/wires.py
@@ -920,6 +920,8 @@ class Wire3D(volmdlr.core.CompositePrimitive3D, WireMixin):
 
         :param side: 'old' or 'new'
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for primitive in self.primitives:
             primitive.frame_mapping_inplace(frame, side)
 
@@ -1751,6 +1753,8 @@ class Contour2D(ContourMixin, Wire2D):
         return Contour2D(self.inverted_primitives())
 
     def invert_inplace(self):
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         self.primitives = self.inverted_primitives()
 
     def random_point_inside(self, include_edge_points: bool = False):
@@ -2580,6 +2584,8 @@ class ClosedPolygon2D(Contour2D, ClosedPolygonMixin):
         :param center: rotation center
         :param angle: rotation angle
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for point in self.points:
             point.rotation_inplace(center, angle)
 
@@ -2599,6 +2605,8 @@ class ClosedPolygon2D(Contour2D, ClosedPolygonMixin):
 
         :param offset: translation vector
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for point in self.points:
             point.translation_inplace(offset)
 
@@ -2606,6 +2614,8 @@ class ClosedPolygon2D(Contour2D, ClosedPolygonMixin):
         return self.__class__([point.frame_mapping(frame, side) for point in self.points])
 
     def frame_mapping_inplace(self, frame: volmdlr.Frame2D, side: str):
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for point in self.points:
             point.frame_mapping_inplace(frame, side)
 
@@ -3960,6 +3970,8 @@ class Circle2D(Contour2D):
         :param center: rotation center
         :param angle: rotation angle
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         self.center.rotation_inplace(center, angle)
 
     def translation(self, offset: volmdlr.Vector2D):
@@ -3977,6 +3989,8 @@ class Circle2D(Contour2D):
 
         :param offset: translation vector
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         self.center.translation_inplace(offset)
 
     def frame_mapping(self, frame: volmdlr.Frame3D, side: str):
@@ -3985,10 +3999,10 @@ class Circle2D(Contour2D):
         side = 'old' or 'new'
         """
         if side == 'old':
-            return Circle2D(frame.old_coordinates(self.center),
+            return Circle2D(frame.local_to_global_coordinates(self.center),
                             self.radius)
         elif side == 'new':
-            return Circle2D(frame.new_coordinates(self.center),
+            return Circle2D(frame.global_to_local_coordinates(self.center),
                             self.radius)
         else:
             raise ValueError('Side should be \'new\' \'old\'')
@@ -3998,10 +4012,12 @@ class Circle2D(Contour2D):
         Changes frame_mapping and the object is updated inplace
         side = 'old' or 'new'
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         if side == 'old':
-            self.center = frame.old_coordinates(self.center)
+            self.center = frame.local_to_global_coordinates(self.center)
         elif side == 'new':
-            self.center = frame.new_coordinates(self.center)
+            self.center = frame.global_to_local_coordinates(self.center)
         else:
             raise ValueError('Side should be \'new\' \'old\'')
 
@@ -4303,7 +4319,7 @@ class Ellipse2D(Contour2D):
                              self.major_dir)
         if side == 'new':
             point_major_dir = self.center + self.major_dir * self.major_axis
-            major_dir = frame.new_coordinates(point_major_dir) - self.center
+            major_dir = frame.global_to_local_coordinates(point_major_dir) - self.center
             return Ellipse2D(self.major_axis, self.minor_axis, frame.global_to_local_coordinates(self.center),
                              major_dir)
         raise ValueError('Side should be \'new\' \'old\'')
@@ -4546,6 +4562,8 @@ class Contour3D(ContourMixin, Wire3D):
         :param axis: rotation axis.
         :param angle: rotation angle.
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for edge in self.primitives:
             edge.rotation_inplace(center, axis, angle)
 
@@ -4566,6 +4584,8 @@ class Contour3D(ContourMixin, Wire3D):
 
         :param offset: translation vector.
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for edge in self.primitives:
             edge.translation_inplace(offset)
 
@@ -4619,6 +4639,8 @@ class Contour3D(ContourMixin, Wire3D):
 
         :param side: 'old' or 'new'
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for edge in self.primitives:
             edge.frame_mapping_inplace(frame, side)
 
@@ -4842,7 +4864,7 @@ class Circle3D(Contour3D):
         :param point: point to calculate abscissa.
         :return: abscissa
         """
-        x, y, _ = self.frame.new_coordinates(point)
+        x, y, _ = self.frame.global_to_local_coordinates(point)
         u1 = x / self.radius
         u2 = y / self.radius
         theta = volmdlr.geometry.sin_cos_angle(u1, u2)
@@ -4872,6 +4894,8 @@ class Circle3D(Contour3D):
         :param axis: rotation axis
         :param angle: rotation angle
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         self.frame.rotation_inplace(center, axis, angle)
 
     def translation(self, offset: volmdlr.Vector3D):
@@ -4889,6 +4913,8 @@ class Circle3D(Contour3D):
 
         :param offset: translation vector
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         self.frame.translation_inplace(offset)
 
     def plot(self, ax=None, color='k', alpha=1., edge_details=False):
@@ -5182,7 +5208,7 @@ class Ellipse3D(Contour3D):
         :param point: point to be verified.
         :return: True is point lies on the Ellipse, False otherwise
         """
-        new_point = self.frame.new_coordinates(point)
+        new_point = self.frame.global_to_local_coordinates(point)
         return math.isclose(new_point.x ** 2 / self.major_axis ** 2 +
                             new_point.y ** 2 / self.minor_axis ** 2, 1, abs_tol=1e-6)
 
@@ -5249,7 +5275,7 @@ class Ellipse3D(Contour3D):
         frame = volmdlr.Frame3D(self.center, self.major_dir,
                                 self.normal.cross(self.major_dir), self.normal)
 
-        p1_new, p2_new = frame.new_coordinates(point1), frame.new_coordinates(point2)
+        p1_new, p2_new = frame.global_to_local_coordinates(point1), frame.global_to_local_coordinates(point2)
 
         theta1 = volmdlr.geometry.sin_cos_angle(p1_new.x / self.major_axis, p1_new.y / self.minor_axis)
 
@@ -5260,7 +5286,7 @@ class Ellipse3D(Contour3D):
         else:
             angle = (theta1 + theta2) / 2
 
-        p3 = frame.old_coordinates(volmdlr.Point3D(self.major_axis * math.cos(angle),
+        p3 = frame.local_to_global_coordinates(volmdlr.Point3D(self.major_axis * math.cos(angle),
                                                    self.minor_axis * math.sin(angle), 0))
 
         return volmdlr.edges.ArcEllipse3D(point1, p3, point2, self.center,
@@ -5289,6 +5315,8 @@ class Ellipse3D(Contour3D):
         :param axis: rotation axis
         :param angle: rotation angle
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         self.center.rotation_inplace(center, axis, angle)
         self.normal.rotation_inplace(center, axis, angle)
         self.major_dir.rotation_inplace(center, axis, angle)
@@ -5311,6 +5339,8 @@ class Ellipse3D(Contour3D):
 
         :param offset: translation vector
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         self.center.translation_inplace(offset)
         self.normal.translation_inplace(offset)
         self.major_dir.translation_inplace(offset)
@@ -5423,6 +5453,8 @@ class ClosedPolygon3D(Contour3D, ClosedPolygonMixin):
         :param axis: rotation axis.
         :param angle: rotation angle.
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for point in self.points:
             point.rotation_inplace(center, axis, angle)
 
@@ -5443,6 +5475,8 @@ class ClosedPolygon3D(Contour3D, ClosedPolygonMixin):
 
         :param offset: translation vector.
         """
+        warnings.warn("'inplace' methods are deprecated. Use a not inplace method instead.", DeprecationWarning)
+
         for point in self.points:
             point.translation_inplace(offset)